fix: Show live widget previews
This change includes some drive-by refactors to break structs into separate files.
