+++ BuildsWidget/BuildsWidget.swift
@@ -25,40 +25,12 @@ import Interact
 
 import BuildsCore
 
-struct Provider: AppIntentTimelineProvider {
-
-    enum Key: String {
-        case summary
-    }
-
-    func placeholder(in context: Context) -> SimpleEntry {
-        SimpleEntry(date: Date(), configuration: ConfigurationAppIntent())
-    }
-
-    func snapshot(for configuration: ConfigurationAppIntent, in context: Context) async -> SimpleEntry {
-        SimpleEntry(date: Date(), configuration: configuration)
-    }
-    
-    func timeline(for configuration: ConfigurationAppIntent, in context: Context) async -> Timeline<SimpleEntry> {
-        var entries: [SimpleEntry] = []
-
-        let defaults = KeyedDefaults<Key>(defaults: UserDefaults(suiteName: "group.uk.co.jbmorley.builds")!)
-        let summary: Summary?  = try? defaults.codable(forKey: .summary)
-
-        let currentDate = Date()
-        let entry = SimpleEntry(date: currentDate, summary: summary, configuration: configuration)
-        entries.append(entry)
-
-        return Timeline(entries: entries, policy: .atEnd)
-    }
-}
-
 struct SimpleEntry: TimelineEntry {
     let date: Date
-    let summary: Summary?
+    let summary: Summary
     let configuration: ConfigurationAppIntent
 
-    init(date: Date, summary: Summary? = nil, configuration: ConfigurationAppIntent) {
+    init(date: Date = Date(), summary: Summary = Summary(), configuration: ConfigurationAppIntent) {
         self.date = date
         self.summary = summary
         self.configuration = configuration
@@ -73,12 +45,12 @@ struct BuildsWidgetEntryView : View {
         VStack(alignment: .leading) {
             HStack {
                 Spacer()
-                Image(systemName: entry.summary?.status.systemImage ?? "questionmark")
+                Image(systemName: entry.summary.status.systemImage)
                     .imageScale(.large)
             }
             Spacer()
-            Text("\(entry.summary?.count ?? 0) Workflows")
-            if let date = entry.summary?.date {
+            Text("\(entry.summary.count) Workflows")
+            if let date = entry.summary.date {
                 Text(date, format: .relative(presentation: .named))
                     .font(.subheadline)
                     .opacity(0.6)
@@ -89,7 +61,7 @@ struct BuildsWidgetEntryView : View {
             }
         }
         .widgetAccentable()
-        .containerBackground(entry.summary?.status.color ?? .pink, for: .widget)
+        .containerBackground(entry.summary.status.color, for: .widget)
     }
 }
 
@@ -100,20 +72,7 @@ struct BuildsWidget: Widget {
         AppIntentConfiguration(kind: kind, intent: ConfigurationAppIntent.self, provider: Provider()) { entry in
             BuildsWidgetEntryView(entry: entry)
         }
-    }
-}
-
-extension ConfigurationAppIntent {
-    fileprivate static var smiley: ConfigurationAppIntent {
-        let intent = ConfigurationAppIntent()
-        intent.favoriteEmoji = "😀"
-        return intent
-    }
-    
-    fileprivate static var starEyes: ConfigurationAppIntent {
-        let intent = ConfigurationAppIntent()
-        intent.favoriteEmoji = "🤩"
-        return intent
+        .configurationDisplayName("All Workflows")
     }
 }
